requirements.txt
opencv-python
Pillow
pytesseract
streamlit
